Added better support for future partial put objects
Updated the schema and the DB writing logic to accept NULL values. Prior to this, if a user wanted to have an empty cell in their table, the server wouldnt write it to the DB because it was only writing defined values. I considered instead using an empty string rather than NULL but the DB already defaults to NULL on empty values, so why not just reuse that to make the logic cleaner. In the client's current state, it resends the entire row on PUT requests. However this should allow the server to accept partials in the server. So if a user wants to clear the 'Company' cell, the client can specify "company": null, or if they want to leave it alone, just no specify it.