Added tab editing and saving
Implemented the ability to change a tab's name and color in the application. Also added batched saving (similar to in table) for tab changes.
I also had to seperate some files (tabs and tab) to fix an issue where components would comlpetely remount on state change rather than just rerendering.
The PPAPP is not finished, just a placeholder to display the multitable for now.